+++ NEWS.md
@@ -1,8 +1,9 @@
 # log history of ech package development
 
-## ech v0.1.3 (in progress)
+## ech v0.1.3 (in progress) 
 * Add to dictionary variables of ech 2021
 * Add function organize_e193
+* minor changes. fix dplyr::between error
 
 ## ech v0.1.2.0
 *minor changes. fix organize_names